+++ package.json
@@ -17,6 +17,7 @@
     "react-redux": "^8.0.5",
     "react-router-dom": "^6.4.5",
     "react-scripts": "5.0.1",
+    "reactjs-popup": "^2.0.5",
     "typescript": "^4.9.4",
     "web-vitals": "^2.1.4"
   },
