+++ src/App.tsx
@@ -5,7 +5,6 @@ import HomePage from "./pages/HomePage";
 import Layout from "./components/Layout";
 import NotFoundPage from "./pages/NotFoundPage";
 import ProductDetailsPage from "./pages/ProductDetailsPage";
-import ModalPopupPage from "./pages/ModalPopupPage";
 
 import "./styles/colors.css";
 import "./styles/Footer.css";
@@ -17,7 +16,6 @@ function App() {
         <Route path="/" element={<Layout />}>
           <Route index element={<HomePage />} />
           <Route path="/product-card/:id" element={<ProductDetailsPage />} />
-          <Route path="/ModalPopupPage" element={<ModalPopupPage />} />
           <Route path="/cart" element={<Cart />} />
           <Route path="*" element={<NotFoundPage />} />
         </Route>
