+++ src/styles/Popup.css
@@ -1,8 +1,9 @@
 .modal-popup {
   display: flex;
-  justify-content: center;
   align-items: center;
   flex-direction: column;
+  justify-content: center;
+  background-color: aliceblue;
   border: 1px solid var(--color-orange);
 }
