+++ src/styles/ProductDetails.css
@@ -1,5 +1,6 @@
 .product-details {
   display: flex;
+  position: relative;
   flex-direction: row;
   justify-content: center;
   border: solid 1px var(--color-light-gray);
